Move event synchronization classes into statistics.event subpackage (CF-453)
