+++ ta4j-core/src/main/java/org/ta4j/core/indicators/statistics/event/EventSignal.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.statistics;
+package org.ta4j.core.indicators.statistics.event;
 
 import org.ta4j.core.BarSeries;
