+++ ta4j-core/src/main/java/org/ta4j/core/indicators/statistics/event/EventSignals.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.statistics;
+package org.ta4j.core.indicators.statistics.event;
 
 import java.util.Objects;
 import java.util.function.IntPredicate;
