+++ ta4j-core/src/main/java/org/ta4j/core/indicators/statistics/package-info.java
@@ -21,8 +21,10 @@
  * {@link org.ta4j.core.indicators.statistics.MutualInformationIndicator}, and
  * {@link org.ta4j.core.indicators.statistics.RegimeSegmentedCorrelationIndicator}.
  * Sparse near-coincident Boolean event streams are scored with
- * {@link org.ta4j.core.indicators.statistics.EventSynchronizationIndicator}, a
- * rolling F1 scorer over deterministic one-to-one event matching.
+ * {@link org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator},
+ * a rolling F1 scorer over deterministic one-to-one event matching; the
+ * event-dependence machinery lives in the
+ * {@link org.ta4j.core.indicators.statistics.event} subpackage.
  * </p>
  */
 package org.ta4j.core.indicators.statistics;
