+++ ta4j-core/src/test/java/org/ta4j/core/indicators/statistics/event/EventSynchronizationBenchmarkTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.indicators.statistics;
+package org.ta4j.core.indicators.statistics.event;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
