+++ ta4j-examples/src/main/java/ta4jexamples/analysis/EventSynchronizationExample.java
@@ -10,9 +10,9 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.indicators.statistics.EventSynchronizationIndicator;
-import org.ta4j.core.indicators.statistics.EventSynchronizationIndicator.Result;
-import org.ta4j.core.indicators.statistics.EventSynchronizationIndicator.Result.Match;
+import org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator;
+import org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator.Result;
+import org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator.Result.Match;
 import org.ta4j.core.indicators.NetMomentumIndicator;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.indicators.helpers.ConstantIndicator;
