+++ ta4j-examples/src/test/java/ta4jexamples/analysis/EventSynchronizationExampleTest.java
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-import org.ta4j.core.indicators.statistics.EventSynchronizationIndicator.Result;
+import org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator.Result;
 
 /**
  * Companion regression check for {@link EventSynchronizationExample}.
@@ -33,7 +33,7 @@ class EventSynchronizationExampleTest {
         for (Result result : new Result[] { demo.swingHighs(), demo.swingLows() }) {
             assertEquals(result.predictedCount(), result.matchedCount() + result.unmatchedPredictedIndexes().size());
             assertEquals(result.referenceCount(), result.matchedCount() + result.unmatchedReferenceIndexes().size());
-            for (org.ta4j.core.indicators.statistics.EventSynchronizationIndicator.Result.Match match : result
+            for (org.ta4j.core.indicators.statistics.event.EventSynchronizationIndicator.Result.Match match : result
                     .matches()) {
                 assertTrue(match.offsetBars() >= -12 && match.offsetBars() <= 12,
                         "offset outside the demo tolerance window: " + match);
